// src/enums.ts
export enum EmploymentRate {
  NotApplicable = '6-01', // neither employment nor housing is present
  Full = '8-93', // employment is 95% or higher
  High = '5-03', // employment is 65% or higher
  Average = '1-51', // employment is 35% or higher
  Low = '6-56', // employment is 5% or higher
  None = '3-24', // employment is lower than 5%
}